refactor : 불필요한 BookManager 인터페이스를 클래스로 변경
pr#30
