--- src/test/java/com/dev_course/book/ListBookManagerTest.java
@@ -1,8 +1,0 @@
-package com.dev_course.book;
-
-class ListBookManagerTest extends BookManagerTest {
-    @Override
-    protected BookManager getBookManager() {
-        return new ListBookManager();
-    }
-}
